+++ src/extension/qixfs/entry/index.ts
