+++ src/extension/qixfs/index.ts
@@ -1,1 +1,1 @@
-export * from "./src/utils";
+export * from "./utils";
