+++ src/extension/qixfs/model/src/workspace-folder.ts
