+++ src/tsconfig.json
@@ -17,7 +17,8 @@
       "@extension/enigma/*": ["./extension/enigma/src/*"],
       "@qixfs-entry": ["./extension/qixfs-entry"],
       "@qixfs/*": ["./extension/qixfs/src/*"],
-      "@qixfs/utils/*": ["./extension/qixfs/src/utils/src/*"],
+      "@qixfs/utils": ["extension/qixfs/utils"],
+      "@qixfs/entry": ["extension/qixfs/entry"],
       "@settings/utils" : ["extension/settings/utils"],
       "@settings/model/*" : ["extension/settings/model/*"],
       "@settings/webview/*" : ["extension/settings/webview/*"],
